+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.4
+
+* now getImage TinyMCE plugin doesn't load automtically anymore, if EditorFormats addon is active
+* added sidebar to posts (only published field for now)
+* added lexy renderer settings to GUI
+* fixed typo in Singleton links (custom navigation)
+
 ## 0.1.3
 
 * fixed: settings page didn't apply changes, if config was an empty array